+++ Classlib/tests.cc
@@ -7,7 +7,7 @@
 TestTable test_table[] = {
     {"clString", &clString::test_rigorous, &clString::test_performance},
     {"Queue", Queue<int>::test_rigorous, 0},
-    {"Array1", Array1<int>::test_rigorous, 0},
+    {"Array1", Array1<float>::test_rigorous, 0},
     {0,0,0}
 };
